split armor functions out into spearate script file
